feat: divide js-code into modules
